refactor(session-pr2): extract RouterHistoryBuffer from ChatSession
Extracts 3 history/SP methods + 3 module-level helpers from ChatSession
into RouterHistoryBuffer (session.py refactor PR-2).

- New: `src/reyn/chat/services/router_history_buffer.py`
  - `build_history()` — token-budget router-view slicer (#1128 step 3)
  - `decompose_history_for_retry()` — head/raw_middle/tail for retry_loop
  - `build_system_prompt()` — assembles router SP string
  - `_serialise_turn()` / `_resolve_budgets()` — private helpers
  - `_is_force_close_consolidation`, `_materialise_path_ref_content`,
    `_read_pathref_image` — moved from session.py module scope
- `session.py`: one-line forwarding wrappers; `_history_buffer` created
  before CompactionController (bootstrap order: SP provider is called
  during CompactionEngine.recompute_budgets() at construction time —
  `_history_buffer` must exist; `compaction_controller` is None during
  that call, patched immediately after); `history_fn` in `_budget_advisor`
  updated to `self._history_buffer.build_history`
- `services/__init__.py`: export `RouterHistoryBuffer`
- Tests updated to import helpers from their new canonical location:
  `test_force_close_covers_slice_1092.py`, `test_path_ref_materialisation.py`,
  `test_user_image_input.py` (`_make_history_builder` harness wired)

SP byte-identity scaffold (1 test) ran green — forwarding wrapper and
direct call produce identical bytes — then deleted per testing.ja.md.
