+++ src/reyn/chat/services/__init__.py
@@ -9,6 +9,7 @@ from reyn.chat.services.intervention_handler import InterventionHandler
 from reyn.chat.services.intervention_registry import InterventionRegistry
 from reyn.chat.services.memory_service import MemoryService
 from reyn.chat.services.plan_runner import PlanRunner
+from reyn.chat.services.router_history_buffer import RouterHistoryBuffer
 from reyn.chat.services.router_host_adapter import RouterHostAdapter
 from reyn.chat.services.skill_runner import SkillRunner
 from reyn.chat.services.snapshot_journal import SnapshotJournal
@@ -32,6 +33,7 @@ __all__ = [
     "InterventionRegistry",
     "MemoryService",
     "PlanRunner",
+    "RouterHistoryBuffer",
     "RouterHostAdapter",
     "SkillRunner",
     "SnapshotJournal",
